profile: add build tag matprofile to allow for conditional inclusion of custom profile
