+++ cmd/mjpeg-streamer/main.go
@@ -22,6 +22,8 @@ import (
 	"net/http"
 	"os"
 
+	_ "net/http/pprof"
+
 	"github.com/hybridgroup/mjpeg"
 	"gocv.io/x/gocv"
 )
